refactor(delegation): create x-chain param type
